Rename debug-tests utils files
