--- tests/debug-tests/test_utils.py
@@ -1,1 +1,0 @@
-../utils.py
